refactor(ui): remove dead filledButtonStyles and consolidate CONFIRM_CARD_HORIZONTAL_DECORATION imports
Remove the unused `filledButtonStyles` export from commonViewStyles (no consumers).

Complete the `CONFIRM_CARD_HORIZONTAL_DECORATION` migration: all six remaining
consumers (BashApproval, PlanApproval, AgentProposal, UserQuestion, ExternalInquiry,
confirmCardRowsBudget) now import from `../ui/theme` directly. The re-export shim
in ConfirmCard is no longer needed and is removed.
