+++ packages/cli/src/chat/tui/modals/BashApproval.tsx
@@ -3,7 +3,8 @@ import { Box, Text, useWindowSize } from 'ink';
 
 import type { BashPermission } from '@shared/schemas';
 
-import { ConfirmCard, CONFIRM_CARD_HORIZONTAL_DECORATION } from './ConfirmCard';
+import { ConfirmCard } from './ConfirmCard';
+import { CONFIRM_CARD_HORIZONTAL_DECORATION } from '../ui/theme';
 import { confirmCardContentRowsBudget } from './confirmCardRowsBudget';
 import { wrapAnsiToWidth } from '../render/ansiWrap';
 import {
