+++ packages/cli/src/chat/tui/modals/UserQuestion.tsx
@@ -14,7 +14,7 @@ import {
   USER_QUESTION_SKIPPED_FEEDBACK,
   userQuestionDecision,
 } from './UserQuestionState';
-import { CONFIRM_CARD_HORIZONTAL_DECORATION } from './ConfirmCard';
+import { CONFIRM_CARD_HORIZONTAL_DECORATION } from '../ui/theme';
 import { BaseTextInput } from '../input/BaseTextInput';
 import { isEscapeInput, isPlainReturnInput } from '../input/inputKeys';
 import { wrapAnsiToWidth } from '../render/ansiWrap';
